Fix integer overflow in output
git-svn-id: svn+ssh://svn.umic-mesh.net/umic-mesh/tools/flowgrind/branches/flowgrind-distributed@4893 3d4b872d-e00e-0410-a2ce-a5c2cda4747c

+++ src/flowgrind.c
@@ -640,7 +640,7 @@ void log_output(const char *msg)
 void print_tcp_report_line(char hash, int id,
 		int type, /* 0 source 1 destination */
 		double time1, double time2,
-		long bytes_written, long bytes_read,
+		long long bytes_written, long long bytes_read,
 		long read_reply_blocks,  double min_rtt,
 		double tot_rtt, double max_rtt, double min_iat,
 		double tot_iat, double max_iat,
